Bug 424836 XViewerEditAdapter extensions
the CellEditDescriptor has a new constructor with the additional boolean
"fitInCell" - default is ture. if this boolean is false the created
control doesn't fit in the klicked cell. it will just start at the top
left corner of the cell.

the XViewerConverter getInput method returns object for object changes
with new references (null allowed). it also have the new methode isValid
which decides if the selected object is allowed to edit at the clicked
cell (same as ICellModifier.canModify).

the XViewerEditAdapter is prepared to extend from it. the
XViewerMultiEditAdapter (which is extended from the XViewerEditAdapter)
can handle multi select and edit.

XViewer.dispose updated (menuManager disposed; display filter removed)

Change-Id: I5670e0fb996d30f02abf36119b4ca494a5466afc
Signed-off-by: Juergen Reichl <reichljuergen@googlemail.com>