Bug 419928 - GridTableViewer and Grid performance improvement
There is no more collection within the GridItem , this greatly improves
the creation and management of data.
Everything is based on DataVisualizer . This is a centralized component
that is allowed to set and get information such as: what is my fonts?
what is my background? what is my text ?
I implemented three types of DataVisualizer :
1 ) ColumnRowBigDataVisualizer . This simply saves the data from column
/ row in Map < Integer , List Of T >
2 ) RowColumnBigDataVisualizer , as above, but save rows / columns
3 ) the real news, the AdaptedDataVisualizer that works as a
LabelProvider . Thanks to AdaptedDataVisualizer Shot down the memory
usage
I fixed some bugs for memory leaks and improve performance of Grid
OnPaint method ..
A change that I could not give to you is the removal of the IsVisible
method and the reimplementation of all the calculations of the first
index to be seen, the performance increase greatly.

Change-Id: Iff715098d98f865f5b22a82cf8f11a7690426e08

Signed-off-by: mpaturzo <caosmpz@yahoo.it>