Bug 430106 - Data held by GridItem are incoherent
I fixed the bug and I made some tests using snippets. 
I have increased the performance of the Grid, further reducing the size
of the memory and I have extended the concept of virtual data visualizer
also to the trees. 
I ran all the snippets and they worked. 

But there is a problem: i had to change the interface DataVisualizer. Is
a major change, i had to replace the row index parameter with the
reference to GridItem.


Change-Id: I71093eec1d816364f82535fbb78db790fc2f2f1a
Signed-off-by: mpaturzo <caosmpz@yahoo.it>