Bug 460119 ArrayIndexOutOfBoundsException in XYGraph when graphs are drawn
Modified the getStraightLineIntersection method in Trace.java to handle the case where the line crosses
1 point on the x axes, but two points on the y axes.  In this case one of the y axes points will be the
same as the x axes point so now add checks when adding the y axis points so that new points are added only
if they are not already in the array.

Tested by running against the code in the bug report (460119) previously this lead to an ArrayIndexOutOfBoundsException, with the new code it does not.

Signed-off-by: Alex Clayton <Alex_Clayton_2000@yahoo.com>