Bug 467676 - Allow use of custom axes in XYGraph
Introduce an Axis factory to XYGraph that can be used by the XYGraph constructor to create its primary axis.  This will allow users to specify custom axes to be used
for the primary axes.

A new constructor is added to XYGraph that takes a AxesFactory as an argument.  It will use this factory to create the primary axes.  The old constructor now delegates to this
new constructor passing it a default factory, this default factory mimics the behaviour for axes creation that was in the old code.  This should mean that existing code still works as it used to.

Tested by running the examples they still work as expected.
Tested by using new code in my own work to create custom axes.

Change-Id: If11db5da85e6afc11ee3614fff8011a1f8dce3af
Signed-off-by: Alex Clayton <Alex_Clayton_2000@yahoo.com>