Bug 477896 - Please include setup for project Nebula
I've added an Eclipse Ini task to redirect the web-hosted setup to the
version in the git clone so that you can edit the local setup and test
it locally before committing it.

Also, unless the source locator looks at nested projects, it only finds
the root projects and imports no projects into the workspace.

Change-Id: Iadb5dfa1ccdb7cf36f00cd846cfc69a457ba8318
Signed-off-by: Ed Merks <ed.merks@gmail.com>