[383589] - Verify text modification event can be processed multiple times
Due to the way listeners are handles by SWT (an array of listeners), it
is possible that the verify method could be called twice for the same
SWT.Verify event. At the bottom of the  verify method in CDateTime the
updateText() method is called which in turns calls updateText(boolean
async).  At the bottom of the method a runnable is created which
modifies the text control.  The verify listener is removed, the text
control setText is called, and then the listener is added back.  When
the listener is added back, it will be added to the end of the array of
listeners that are currently being called, thus being recalled.  Adding
if (e.doit == false) { return; } in the verify listener fixes this
issue.

Bug: https://bugs.eclipse.org/bugs/show_bug.cgi?id=383589

Change-Id: Ib5867d01eb28132adadf023bab52635090aac58c
Signed-off-by: Doug Showell <dats07@gmail.com>