Bug 485173 - Null pointer exception in CDateTimeObservableValue.java
The CDateTime selection will be null if the user hasn't entered any
value.  The text "<choose date>" appears when the value has been set to
null.  However, the CDateTimeObservableValue class doesn't handle the
case of 'getSelection()' returning null (an advertised possible return
value - see the JavaDoc for CDateTime.getSelection()).

In addition, the fix for Bug 403355 allows generation a property change
event when the current and new values are the same.  

The attached patch fixes both these problems.

Change-Id: I256ba944aa982396467ea5b8bc0423aaab4ad899
Signed-off-by: Wim Jongman <wim.jongman@remainsoftware.com>
Signed-off-by: Glenn Burkhardt <glenn.burkhardt@utas.utc.com>