Bug 288164 - Month selector in date dropdown has two March's.
for the month dropdown, it suffices to use the 1st day of month
in the calendar instance to fix the two March's.

The changing month mentioned in the comment can be fixed by setting the
selected date to the minimum of the actual maximum date for the month
and the selected date.

Signed-off-by: John Janus <j.janus@lighthouse-it.de>