Bug 385585 - Tablecombo popup displays a blank table when ..
.. the Shell containing the tablecombo widget is detached.

The problem was that the popup shell containing the table had a parent
shell which was the shell from the Composite that the TableCombo was
drawn on.

In other words, the Composite of the View and the popup Shell had the
same Shell as parent. There is a setParent(Composite) method but
that does not work for Shells.

The Shell.reparent(Composite) does not throw an Exception but returns
an int signaling that the reparent does not work. This was not checked.

Then the code contained a bunch of code to rebuild the popup Shell when
the parent shell was changed. This could normally work only it also
created a new, empty, Table because the Table has 'popup'
as a parent so the TableItems were also disposed and not rebuild.

The solution was to give the 'popup' a new Shell directly based on
Display and not on the passed parent.

I removed the 'try to rebuild' code and applied code formatting.
All tests work fine.

Change-Id: Id688354194ba81b9e95a2f8fbf651fb61a5cdd92
Signed-off-by: Wim Jongman <wim.jongman@remainsoftware.com>