Add support to the slider to format the value
Currently the slider always show the raw number, but often this is not
very suitable  for the user e.g. if the Number represents a unit of
measure or should be formated with a number format.

This adds a new method NebulaSlider#setLabelFormatProvider that allows
to customize the formating of the raw number.