Do not paint over the borders of the widget
Currently RectangleGroupStrategy sets the clipping to the maximum area
the control will need. But in case where the widget size is larger than
the area to paint (for example inside a ScrolledComposite) this will
result in painting over other widgets (at least on linux).

This now do two things:

 1) save the original clipping region and recover it afterwards
 2) intersect the calculated clipping region with the original to
possibly shrink it to the are that is valid for repaint.